Remidner for tomorow
Each unit is going to use interfaces for updating different behaviours. This will send the required data to update the unit. The data sent will mostly be for aethestics e.g. the enemies postion so it can aim at the target. The attack interface with also include the IFire interface to indicate when it should fire. I'll probably change this but we will see other ideas include having classes attached to each unit similar to DOTS where they just hold data. though they would also contain some logic for animations to use the data passed in.