Attacks onyl check LOS when aiming at target and finding new enemies.
Slight FPS improvement I believe. Some more work needs to go in checking for LOS only when the position of the enemy changes. As when the tank is aiming at the enemy it then starts to check every frame. Some work also needs to be done on colliders and the auto rotate feature as this is causing some issues especially on oddly shaped surfaces. Tempted to just leave each tank with a box collider encompasing the entire tank. Then the raytarget can be left in the exact middle of the unit. Which would be easier for units and yeah. Another potential issue is a target being partially exposed, for now this will be left and the map will be designed so this is less likely to happen.