Added new init for some things
Stuff were trying to access array that weren't initialsied yet. SO made another Custom init that gets ran before the module get's started. Might add a check for the module that rely on update as they may access stuff that isn't initlialised yet too. Currently they are not though as Update is ran after the Units are all spawned and StartModuleManager is ran